+++ barelymusician/internal/message_queue_test.cpp
@@ -18,39 +18,39 @@ namespace {
 // Tests that adding a single message is queued as expected.
 TEST(MessageQueueTest, AddSingleMessage) {
   MessageQueue messages;
-  EXPECT_THAT(messages.GetNext(0.0), IsNull());
-  EXPECT_THAT(messages.GetNext(1.0), IsNull());
-  EXPECT_THAT(messages.GetNext(10.0), IsNull());
+  EXPECT_THAT(messages.GetNext(0), IsNull());
+  EXPECT_THAT(messages.GetNext(1), IsNull());
+  EXPECT_THAT(messages.GetNext(10), IsNull());
 
-  messages.Add(1.0, NoteOffMessage{5.0});
-  EXPECT_THAT(messages.GetNext(0.0), IsNull());
-  EXPECT_THAT(messages.GetNext(1.0), IsNull());
+  messages.Add(1, NoteOffMessage{5});
+  EXPECT_THAT(messages.GetNext(0), IsNull());
+  EXPECT_THAT(messages.GetNext(1), IsNull());
   EXPECT_THAT(
-      messages.GetNext(10.0),
+      messages.GetNext(10),
       AllOf(NotNull(),
-            Pointee(Pair(1.0, VariantWith<NoteOffMessage>(Field(&NoteOffMessage::pitch, 5.0))))));
+            Pointee(Pair(1, VariantWith<NoteOffMessage>(Field(&NoteOffMessage::pitch, 5.0))))));
 
   // Message is already returned.
-  EXPECT_THAT(messages.GetNext(10.0), IsNull());
+  EXPECT_THAT(messages.GetNext(10), IsNull());
 }
 
 // Tests that adding multiple messages are queued as expected.
 TEST(MessageQueueTest, AddMultipleMessages) {
   MessageQueue messages;
-  EXPECT_THAT(messages.GetNext(10.0), IsNull());
+  EXPECT_THAT(messages.GetNext(10), IsNull());
 
-  for (double i = 0.0; i < 10.0; ++i) {
-    messages.Add(i, NoteOffMessage{i});
+  for (int i = 0; i < 10; ++i) {
+    messages.Add(i, NoteOffMessage{static_cast<double>(i)});
   }
-  for (double i = 0.0; i < 10.0; ++i) {
+  for (int i = 0; i < 10; ++i) {
     EXPECT_THAT(
-        messages.GetNext(10.0),
-        AllOf(NotNull(),
-              Pointee(Pair(i, VariantWith<NoteOffMessage>(Field(&NoteOffMessage::pitch, i))))));
+        messages.GetNext(10),
+        AllOf(NotNull(), Pointee(Pair(i, VariantWith<NoteOffMessage>(Field(
+                                             &NoteOffMessage::pitch, static_cast<double>(i)))))));
   }
 
   // All messages are already returned.
-  EXPECT_THAT(messages.GetNext(10.0), IsNull());
+  EXPECT_THAT(messages.GetNext(10), IsNull());
 }
 
 // TODO(#97): Add more tests with concurrency.
